Remove T template parameters from nodes.
Can get the value type from the allocator. The modern C++ way to do this
would be to parameterize the nodes with the pointer type, and then use
pointer traits to get the required types. This would be cleaner, and
potentially generate less code, but could break backwards compatability.
Might make it optional?