Hoist everything that needs the bucket policy.
Can probably merge these base classes. Since maps have a const key, sets
are very unlikely to have the same value type as maps, so there's not
much point in having separate classes - if two value_types are the same,
it's very likely that the iterator and bucket policies will be the same.