Always use const key_type internally
In maps/multimaps I had to keep around both key_type and value_type in
order to get the constness of key_type right. Not really needed, as the
internals don't need to know if key_type is const, so rename internal
key_type to key_type2 to make it clear it's not the same type. Might use
a better name later, might not.