Construct the extra bucket with value.
There's an extra bucket at the end of the buckets which contains a
pointer to the start of the node linked list. Was default constructing
it, then assigning, but now construct with its initial value.

Don't actually need to increment `constructed` at this point as nothing
more can throw, but that might change in the future.